Resolviendo bug de los botones paypal y haciendo primer refactoring
