+++ src/Controllers/constante.js
@@ -23,13 +23,7 @@ const ButtonLogOut = () => {
   return (
     <>
 
-        <Link to='/donar' style={{ textDecoration: "none" }}>
-          <Button sx={{ color: "#7165A0;", "&:hover": { color: "#F50057" } }}>
-            <FavoriteIcon sx={{ marginRight: "10px", fontSize: "18px" }} />
-            Danos Amor $$
-            <FavoriteIcon sx={{ marginRight: "10px", fontSize: "18px" }} />
-          </Button>
-        </Link>,
+        
       <WelcomeUserMsg>
         <h5>Bienvenido</h5>
         <h4>{user.full_name}</h4>
